♻️ folder경로 접근시 folder/0으로 리다이렉트하도록 수정
